# use `memset_patternN` in `Buffer.fill`